HAPP-1332: Create FetchMedicationsUseCase
